+++ webview-ui/src/App.tsx
@@ -103,9 +103,7 @@ const AppContent = () => {
 	}
 
 	return (
-		<div className="flex h-screen w-full flex-col bg-[#1e1e1e]">
-			{/* Fetch Coder Chat Panel - Right Side of VSCode */}
-			{/* VSCode Terminal/Editor runs natively on the left */}
+		<div className="fetch-coder-webview-root flex h-screen w-full flex-col">
 			{showSettings && (
 				<SettingsView
 					onDone={hideSettings}